--- cudax/include/cuda/experimental/__multi_gpu/algorithm/sort/bracket.h
@@ -1,49 +1,0 @@
-// -*- C++ -*-
-//===----------------------------------------------------------------------===//
-//
-// Part of CUDA Experimental in CUDA C++ Core Libraries,
-// under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES.
-//
-//===----------------------------------------------------------------------===//
-
-#ifndef _CUDA_EXPERIMENTAL___MULTI_GPU_ALGORITHM_SORT_BRACKET_H
-#define _CUDA_EXPERIMENTAL___MULTI_GPU_ALGORITHM_SORT_BRACKET_H
-
-#include <cuda/std/detail/__config>
-
-#if defined(_CCCL_IMPLICIT_SYSTEM_HEADER_GCC)
-#  pragma GCC system_header
-#elif defined(_CCCL_IMPLICIT_SYSTEM_HEADER_CLANG)
-#  pragma clang system_header
-#elif defined(_CCCL_IMPLICIT_SYSTEM_HEADER_MSVC)
-#  pragma system_header
-#endif // no system header
-
-#include <cuda/std/__optional/optional.h>
-#include <cuda/std/cstdint>
-
-#include <cuda/std/__cccl/prologue.h>
-
-// NOLINTBEGIN(bugprone-reserved-identifier)
-
-namespace cuda::experimental::__detail
-{
-// Persistent per-splitter bracket. rank = global rank of `key`.
-// L: largest key proven to sit BELOW the ideal rank Ni/p.
-// U: smallest key proven to sit ABOVE  the ideal rank Ni/p.
-template <class _Tp>
-struct __bracket
-{
-  ::cuda::std::uint64_t __rank; // < global rank of the key
-  ::cuda::std::optional<_Tp> __key; // < the key, if found. If nullopt means either +/- inf
-};
-} // namespace cuda::experimental::__detail
-
-// NOLINTEND(bugprone-reserved-identifier)
-
-#include <cuda/std/__cccl/epilogue.h>
-
-#endif // _CUDA_EXPERIMENTAL___MULTI_GPU_ALGORITHM_SORT_BRACKET_H
